requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dotenv>=0.21.0
openai>=0.27.0
langchain>=0.1.20
langchain-community>=0.0.27
langgraph==0.6.4
pyodbc==5.2.0
openpyxl==3.1.5
xlsxwriter==3.2.5
psutil==7.0.0
tabulate==0.9.0  